od_zero_shot/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "od_zero_shot"
version = "0.1.0"
description = "极简/样本高效的零样本城市 OD 网络生成基线工程"
readme = "README.md"
requires-python = ">=3.13"
authors = [
  { name = "Codex" }
]
dependencies = [
  "numpy>=2.0",
  "scipy>=1.13",
  "pandas>=2.2",
  "matplotlib>=3.8",
  "PyYAML>=6.0",
  "scikit-learn>=1.5",
  "torch>=2.8"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
